# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShilpSetu API",
    "description": "One photo. One spoken sentence. A fair-priced, trust-verified listing — no typing, no middleman, no English required. · SIH 2026 · PS ID SIH26090 · Team HACKER LOBBY",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/operators/artisans": {
      "get": {
        "tags": [
          "operators"
        ],
        "summary": "My Artisans",
        "operationId": "my_artisans_api_v1_operators_artisans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "operators"
        ],
        "summary": "Onboard Artisan",
        "description": "Onboard an artisan who may not own a smartphone. The artisan's spoken consent is recorded on the kiosk.",
        "operationId": "onboard_artisan_api_v1_operators_artisans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorArtisanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/verification-queue": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Verification Queue",
        "operationId": "verification_queue_api_v1_admin_verification_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/artisans/{artisan_id}/verify": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Verify Artisan",
        "operationId": "verify_artisan_api_v1_admin_artisans__artisan_id__verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artisan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artisan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyArtisanIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/artisans": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "All Artisans",
        "operationId": "all_artisans_api_v1_admin_artisans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/listings": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Listings",
        "operationId": "listings_api_v1_admin_listings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/listings/{product_id}/moderate": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Moderate",
        "operationId": "moderate_api_v1_admin_listings__product_id__moderate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/certificates/{cert_id}/revoke": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Revoke",
        "operationId": "revoke_api_v1_admin_certificates__cert_id__revoke_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/analytics": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Analytics",
        "operationId": "analytics_api_v1_admin_analytics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/analytics/export.csv": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Analytics Export",
        "description": "Anonymised, aggregated collective analytics (a revenue stream for govt / bulk buyers): state x craft.",
        "operationId": "analytics_export_api_v1_admin_analytics_export_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/operators": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Operators",
        "operationId": "operators_api_v1_admin_operators_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/settings": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Settings ",
        "operationId": "get_settings__api_v1_admin_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Put Settings",
        "operationId": "put_settings_api_v1_admin_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/schemes": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Schemes",
        "operationId": "schemes_api_v1_admin_schemes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Add Scheme",
        "operationId": "add_scheme_api_v1_admin_schemes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemeLinkIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/ondc": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Ondc Status",
        "operationId": "ondc_status_api_v1_admin_ondc_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/ondc/{product_id}/retry": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Ondc Retry",
        "operationId": "ondc_retry_api_v1_admin_ondc__product_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/ondc/sync-all": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Ondc Sync All",
        "operationId": "ondc_sync_all_api_v1_admin_ondc_sync_all_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/models": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Models",
        "operationId": "models_api_v1_admin_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "VerifyArtisanIn": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "verified",
              "rejected"
            ],
            "title": "Status"
          },
          "pehchan_verified": {
            "type": "boolean",
            "title": "Pehchan Verified",
            "default": false
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "VerifyArtisanIn"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}